requirements/prod.txt
Django~=3.2.19
djangorestframework~=3.14.0
django-cleanup==7.0.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
sorl-thumbnail==12.9.0
